+++ packages/client/src/redux/store.ts
@@ -1,10 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit'
-import authReducer from './auth/auth.slice'
 import userReducer from './user/user.slice'
 
 const store = configureStore({
   reducer: {
-    auth: authReducer,
     user: userReducer,
   },
 })
